Day 3: Added navigation, also details, splash,loading, forecast screens, also forecast request to server and a few little things

+++ app/src/main/java/com/example/weather/data/utils/Wind.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class Wind(
     @SerialName("speed")
-    val speed:Int,
+    val speed:Float,
     @SerialName("deg")
     val deg:Int
 )
